# sprint5_biblioteca_api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "📚 Biblioteca REST API",
        "description": "API desarrollada por **Jenifer Alvarez** (Sprint 5 - IT Academy).  \r\n *\r\n * Implementa autenticación con **Laravel Passport**, control de acceso por roles  \r\n * y gestión de **Libros**, **Categorías** y **Préstamos**.  \r\n * Respuestas JSON estructuradas para integrarse con frontends o sistemas de IA.",
        "contact": {
            "name": "Jenifer Álvarez",
            "url": "https://github.com/jenifera5"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Autentificación"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Jenifer"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jenifer@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "rol": {
                                        "description": "Opcional: admin o usuario",
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario registrado correctamente"
                                        },
                                        "usuario": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Jenifer"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jenifer@example.com"
                                                },
                                                "rol": {
                                                    "type": "string",
                                                    "example": "admin"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJh..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos o email duplicado"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Autentificación"
                ],
                "summary": "Iniciar sesión de usuario",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jenifer@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "usuario": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Jenifer"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jenifer@example.com"
                                                },
                                                "rol": {
                                                    "type": "string",
                                                    "example": "admin"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJh..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autentificación"
                ],
                "summary": "Cerrar sesión del usuario autenticado",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sesión cerrada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado o token inválido"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        [
            {
                "bearerAuth": []
            }
        ]
    ]
}